+++ src/components/navbar/index.js
@@ -2,7 +2,7 @@ import UserView from "../user-view"
 import Link from "../re-useable/link"
 const Navbar = () => {
   return (
-    <>
+    <div className="container">
       <ul id="slide-out" className="sidenav">
         <li>
           <UserView title="Learn Kannada" email={"kannada@gmail.com"}/>
@@ -11,43 +11,51 @@ const Navbar = () => {
         <ul class="collapsible collapsible-accordion">
           <li>
             <Link className="collapsible-header">
-            Dropdown<i class="material-icons">arrow_drop_down</i></Link>
+           <i class="material-icons">arrow_drop_down</i> Pratice</Link>
             <div class="collapsible-body">
               <ul>
-                <li><Link url="/">First</Link></li>
-                <li><Link url="/">Second</Link></li>
-                <li><Link url="/">Third</Link></li>
-                <li><Link url="/">Fourth</Link></li>
+                <li><Link url="/Vowels" className="sidenav-close">Vowels</Link></li>
+                <li><Link url="/Constents" className="sidenav-close">Constents</Link></li>
+                <li><Link url="/Numbers" className="sidenav-close">Numbers</Link></li>
               </ul>
             </div>
           </li>
         </ul>
       </li>
         <li>
-          <Link url="/">
-            <i className="material-icons">cloud</i>First Link With Icon
+          <Link url="/" className="sidenav-close">
+            Vowels
           </Link>
         </li>
         <li>
-          <Link url="/">Second Link</Link>
+          <Link className="sidenav-close" url="/Constents">Constents</Link>
+        </li>
+        <li>
+          <Link className="sidenav-close" url="/Vowels">Vowels</Link>
+        </li>
+        <li>
+          <Link className="sidenav-close" url="/Numbers">Numbers</Link>
         </li>
         <li>
           <div className="divider"></div>
         </li>
         <li>
-          <Link className="subheader">Subheader</Link>
+          <Link className="sidenav-close" className="subheader">Share</Link>
         </li>
         <li>
-          <Link className="waves-effect" url="/">
-            Third Link With Waves
+          <Link className="sidenav-close subheader">Help</Link>
+        </li>
+        <li>
+          <Link className="sidenav-close subheader" url="/">
+            About
           </Link>
         </li>
 
       </ul>
       <Link url="/" dataTarget="slide-out" className="sidenav-trigger">
-        <i className="material-icons">menu</i>
+        <i className="material-icons" style={{fontSize:"34px"}}>menu</i>
       </Link>
-    </>
+    </div>
   );
 };
